Read (or convert) from TMD file to x3p

TMD files are used in telemetry, specifically, they are a native format used by GelSight to store 3d topographic surface scans.

Usage

tmd_to_x3p(tmd_path, yaml_path = NA, verbose = TRUE)

Arguments

tmd_path

path to TMD file

yaml_path

path to corresponding yaml file with meta information. If set to NA (default), path of the the tmd file will be tried. If set to NULL, meta file will be ignored.

verbose

boolean

Value

x3p file of the scan. Some rudimentary information will be filled in, information of scanning process, and parameter settings need to be added manually.

Details

The algorithm is based on GelSight's MatLab routine readtmd.m published as part of the Github repository gelsightinc/gsmatlab

Examples

#x3p <- tmd_to_x3p("~/Downloads/Sc04.Pl044.Ma4.SB.An80.Pb.DirFo.SizL.tmd") # 
#x3p <- tmd_to_x3p("~/Downloads/Sc04.Pl044.Ma4.SB.An80.Pb.DirFo.SizL.tmd", 
#                   yaml_path="~/Downloads/scan.yaml") #