Read an x3p file into an x3p object

Read file in x3p format. x3p formats describe 3d topological surface according to ISO standard ISO5436 – 2000. x3p files are a container format implemented as a zip archive of a folder consisting of an xml file of meta information and a binary matrix of numeric surface measurements.

Usage

x3p_read(file, size = NA, quiet = T, tmpdir = NULL)

read_x3p(file, size = NA, quiet = T, tmpdir = NULL)

Arguments

file

The file path to the x3p file, or an url to an x3p file

size

size in bytes to use for reading the binary file. If not specified, default is used. Will be overwritten if specified in the xml meta file.

quiet

for url downloads, show download progress?

tmpdir

temporary directory to use to extract the x3p file (default NULL uses tempdir() to set a directory).

Value

x3p object consisting of a list of the surface matrix and the four records as specified in the ISO standard

Examples

logo <- x3p_read(system.file("csafe-logo.x3p", package="x3ptools"))