Read all x3p files in a folder

This function is provided with a path to a folder, it will search recursively for all x3p files within the folder, read the x3p files, and return the results as part of a tibble dataset.

Usage

x3p_read_folder(path)

Arguments

path

character string to a folder

Value

tibble with two variables: source contains the file path to the x3p file, and x3p is a list variable of x3p objects.

Examples

# example code
path <- dirname(system.file("csafe-logo.x3p", package="x3ptools"))
x3p_read_folder(path) # show all x3p files installed with x3ptools
#> # A tibble: 4 × 2
#>   source                                                   x3p          
#>   <chr>                                                    <I<list>>    
#> 1 /home/runner/work/_temp/Library/x3ptools/csafe-logo.x3p  <x3p[,419,1]>
#> 2 /home/runner/work/_temp/Library/x3ptools/pyramid.x3p     <x3p[,5,1]>  
#> 3 /home/runner/work/_temp/Library/x3ptools/sample-land.x3p <x3p[,256,2]>
#> 4 /home/runner/work/_temp/Library/x3ptools/testing.x3p     <x3p[,20,1]>